Publish project package
